Create transpose matix
Initialize a Multidimensional array of  size 3x3.  E.g.- 1 2 3 
4 5 6 
7 8 9 
 
Check if the matrix is symmetric or not. *Few important points to remember: A Square Matrix is said to be symmetric if it is equal to it's transpose. Transpose of a matrix is achieved by exchanging indices of rows and columns. Transpose is only defined for a square matrix.